Rename files for MetricFrame
