+++ fairlearn/metrics/__init__.py
@@ -26,7 +26,7 @@ from ._disparities import (  # noqa: F401
     demographic_parity_ratio,
     equalized_odds_difference,
     equalized_odds_ratio)
-from ._metrics_frame import MetricFrame  # noqa: F401
+from ._metric_frame import MetricFrame  # noqa: F401
 
 from ._extra_metrics import (  # noqa: F401
     true_positive_rate,
