+++ fairlearn/metrics/_disparities.py
@@ -4,7 +4,7 @@
 """Metrics for measuring disparity."""
 
 from ._extra_metrics import selection_rate, true_positive_rate, false_positive_rate
-from ._metrics_frame import MetricFrame
+from ._metric_frame import MetricFrame
 
 
 def demographic_parity_difference(
@@ -40,9 +40,9 @@ def demographic_parity_difference(
         The demographic parity difference
     """
     sel_rate = MetricFrame(selection_rate,
-                            y_true, y_pred,
-                            sensitive_features=sensitive_features,
-                            sample_params={'sample_weight': sample_weight})
+                           y_true, y_pred,
+                           sensitive_features=sensitive_features,
+                           sample_params={'sample_weight': sample_weight})
     result = sel_rate.difference(method='between_groups')['selection_rate']
     return result
 
@@ -80,9 +80,9 @@ def demographic_parity_ratio(
         The demographic parity ratio
     """
     sel_rate = MetricFrame(selection_rate,
-                            y_true, y_pred,
-                            sensitive_features=sensitive_features,
-                            sample_params={'sample_weight': sample_weight})
+                           y_true, y_pred,
+                           sensitive_features=sensitive_features,
+                           sample_params={'sample_weight': sample_weight})
     result = sel_rate.ratio(method='between_groups')['selection_rate']
     return result
 
@@ -172,7 +172,7 @@ def _get_eo_frame(y_true, y_pred, sensitive_features, sample_weight) -> MetricFr
     sw_dict = {'sample_weight': sample_weight}
     sp = {'tpr': sw_dict, 'fpr': sw_dict}
     eo = MetricFrame(fns,
-                      y_true, y_pred,
-                      sensitive_features=sensitive_features,
-                      sample_params=sp)
+                     y_true, y_pred,
+                     sensitive_features=sensitive_features,
+                     sample_params=sp)
     return eo
