+++ fairlearn/metrics/_group_metric_set.py
@@ -13,7 +13,7 @@ from ._extra_metrics import (_balanced_root_mean_squared_error,
                              mean_prediction,
                              selection_rate,
                              true_negative_rate)
-from ._metrics_frame import MetricFrame
+from ._metric_frame import MetricFrame
 from ._input_manipulations import _convert_to_ndarray_and_squeeze
 
 _Y_TRUE = 'trueY'
@@ -161,7 +161,7 @@ def _create_group_metric_set(y_true,
             metric_dict = dict()
             for metric_key, metric_func in function_dict.items():
                 gmr = MetricFrame(metric_func,
-                                   result[_Y_TRUE], prediction, sensitive_features=g[_BIN_VECTOR])
+                                  result[_Y_TRUE], prediction, sensitive_features=g[_BIN_VECTOR])
                 curr_dict = dict()
                 curr_dict[_GLOBAL] = gmr.overall[0]
                 curr_dict[_BINS] = list(gmr.by_group[metric_func.__name__])
